backend/requirements.txt
alembic==1.14.0
arrow==1.3.0
asyncio==3.4.3
attrs==24.3.0
backend==0.2.4.1
chromadb==0.6.2
datasets==3.2.0
docx2txt==0.8
fastapi==0.115.6
google-api-python-client==2.159.0
grpcio==1.69.0
Jinja2==3.1.5
litellm==1.57.5
loguru==0.7.3
numpy==2.2.1
ollama==0.4.5
pandas==2.2.3
pinecone==5.4.2
psycopg2==2.9.10
pydantic==2.10.5
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
PyYAML==6.0.2
py-zerox==0.0.7
qdrant_client==1.12.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
retrying==1.3.4
slack_sdk==3.34.0
SQLAlchemy==2.0.36
supabase==2.11.0
six==1.17.0
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.67.1
weaviate_client==4.10.2
itsdangerous==2.2.0
phidata==2.7.8
youtube_transcript_api==0.6.3
PyGithub==2.5.0